+++ src/common/propertyFieldHeader/PropertyFieldHeader.tsx
@@ -27,10 +27,12 @@ export default class PropertyFieldHeader extends React.Component<IPropertyFieldH
                         {this.props.label}
                     </div>
                     <div className={styles.info}>
+                    {this.props.calloutContent && (
                         <i className={'ms-Icon ms-Icon--Info'} ref={(infoIcon) => { this._infoIcon = infoIcon; }}
-                            onMouseOver={this.props.calloutTrigger === CalloutTriggers.Hover ? this._onInfoIconMouseOver.bind(this) : null}
-                            onMouseOut={this.props.calloutTrigger === CalloutTriggers.Hover ? this._onInfoIconMouseOut.bind(this) : null}
-                            onClick={this.props.calloutTrigger === CalloutTriggers.Click ? this._onInfoIconClick.bind(this) : null}></i>
+                        onMouseOver={this.props.calloutTrigger === CalloutTriggers.Hover ? this._onInfoIconMouseOver.bind(this) : null}
+                        onMouseOut={this.props.calloutTrigger === CalloutTriggers.Hover ? this._onInfoIconMouseOut.bind(this) : null}
+                        onClick={this.props.calloutTrigger === CalloutTriggers.Click ? this._onInfoIconClick.bind(this) : null}></i>
+                    )}
                     </div>
                     {this.state.isCalloutVisible && (
                         <Callout