+++ src/index.ts
@@ -16,3 +16,5 @@ export * from './PropertyFieldLinkWithCallout';
 export * from './PropertyFieldSliderWithCallout';
 export * from './PropertyFieldTextWithCallout';
 export * from './PropertyFieldToggleWithCallout';
+export * from './PropertyPaneWebPartInformation';
+export * from './PropertyPanePropertyEditor';