+++ src/loc/mystrings.d.ts
@@ -1,4 +1,10 @@
 declare interface IPropertyControlStrings {
+
+
+
+  DescriptionLabel: string;
+  MoreInfoLabel: string;
+  
   // PeoplePicker labels
   PeoplePickerSuggestedContacts: string;
   PeoplePickerSuggestedGroups: string;
@@ -83,6 +89,12 @@ declare interface IPropertyControlStrings {
   CollectionDataItemShowErrorsLabel: string;
   CollectionDataItemFieldRequiredLabel: string;
   InvalidUrlError: string;
+
+  // Property Editor 
+  ApplyButtonLabel: string;
+  ImportButtonLabel: string;
+  ExportButtonLabel: string;
+  JsonFileRequiredMessage: string;
 }
 
 declare module 'PropertyControlStrings' {