+++ src/propertyFields/collectionData/collectionNumberField/ICollectionNumberFieldState.ts
@@ -1,3 +1,4 @@
 export interface ICollectionNumberFieldState {
+  value: number;
   errorMessage: string;
 }