+++ src/propertyFields/propertyEditor/index.ts
@@ -1,0 +1,4 @@
+export * from './IPropertyPanePropertyEditor';
+export * from './PropertyPanePropertyEditor';
+export * from './IPropertyPanePropertyEditorHost';
+export * from './PropertyPanePropertyEditorHost';