+++ src/propertyFields/webPartInformation/IPropertyPaneWebPartInformationHost.ts
@@ -1,0 +1,11 @@
+import { IVideoEmbedProperties } from "./IPropertyPaneWebPartInformation";
+
+/**
+ * PropertyFieldColorPickerHost properties interface
+ */
+export interface IPropertyPaneWebPartInformationHostProps {
+	videoProperties?: IVideoEmbedProperties;
+	moreInfoLink?: string;
+	moreInfoLinkTarget?: string;
+	description: string;
+}