+++ src/propertyFields/webPartInformation/index.ts
@@ -1,0 +1,4 @@
+export * from './IPropertyPaneWebPartInformation';
+export * from './PropertyPaneWebPartInformation';
+export * from './IPropertyPaneWebPartInformationHost';
+export * from './PropertyPaneWebPartInformationHost';